restore

+++ database/migrations/2024_10_18_141504_create_photos_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->dateTime( 'approve_date')->nullable();
             $table->unsignedBigInteger( 'buyout_by')->nullable();
             $table->dateTime( 'buyout_date')->nullable();
-            $table->double( 'rate', 8,2)->nullable();
+            $table->decimal( 'rate', 8, 2)->nullable();
             $table->enum( 'status',['pending','approved','declined','selling','buy-out','approve_unselleble'])->default('pending');
             $table->timestamps();
